test: MET-2227 add automation test for transaction overview
